@@ -140,50 +141,37 @@
     "uuid": "^11.0.3",
     "ws": "^8.18.0"
   },
-  "overrides": {
-    "axios": "^1.7.9",
-    "binary-install": {
-      "axios": "^1.7.9"
-    },
-    "wasm-pack": {
-      "axios": "^1.7.9"
-    }
-  },
   "devDependencies": {
-    "@babel/core": "^7.26.0",
-    "@babel/preset-env": "^7.26.0",
-    "@babel/plugin-transform-class-properties": "^7.25.9",
-    "@babel/plugin-transform-private-methods": "^7.25.9",
-    "@babel/plugin-syntax-dynamic-import": "^7.8.3",
-    "@eslint/js": "^9.30.1",
-    "@rollup/plugin-commonjs": "^25.0.0",
-    "@rollup/plugin-node-resolve": "^15.2.0",
+    "@babel/core": "^7.28.3",
+    "@babel/preset-env": "^7.28.3",
+    "@eslint/js": "^9.34.0",
+    "@rollup/plugin-commonjs": "^28.0.6",
+    "@rollup/plugin-node-resolve": "^16.0.1",
     "@rollup/plugin-wasm": "^6.2.0",
-    "@types/node": "^18.19.115",
-    "@types/uuid": "^9.0.0",
+    "@types/node": "^24.3.0",
+    "@types/uuid": "^10.0.0",
     "@types/ws": "^8.5.0",
-    "@typescript-eslint/eslint-plugin": "^8.35.1",
-    "@typescript-eslint/parser": "^8.35.1",
-    "@vitest/ui": "^2.1.9",
-    "babel-jest": "^29.7.0",
-    "core-js": "^3.43.0",
-    "eslint": "^9.30.1",
+    "@typescript-eslint/eslint-plugin": "^8.40.0",
+    "@typescript-eslint/parser": "^8.40.0",
+    "@vitest/ui": "^3.2.4",
+    "babel-jest": "^30.0.5",
+    "core-js": "^3.45.1",
+    "eslint": "^9.34.0",
     "eslint-import-resolver-typescript": "^4.4.4",
     "eslint-plugin-import": "^2.32.0",
     "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-security": "^3.0.1",
-    "jest": "^29.7.0",
+    "jest": "^30.0.5",
     "jest-html-reporters": "^3.1.7",
     "nodemon": "^3.0.0",
     "nyc": "^17.1.0",
-    "rollup": "^3.29.0",
-    "sqlite3": "^5.1.6",
-    "typescript": "^5.2.0",
-    "vitest": "^2.0.0",
+    "rollup": "^4.48.1",
+    "typescript": "^5.9.2",
+    "vitest": "^3.2.4",
     "wasm-opt": "^1.3.0",
     "wasm-pack": "^0.13.1",
-    "webpack": "^5.88.0",
-    "webpack-cli": "^5.1.4"
+    "webpack": "^5.101.3",
+    "webpack-cli": "^6.0.1"
   },
   "optionalDependencies": {
     "worker_threads": "*"
